IR.cpp/hpp renamed
