+++ test/nnfusion/core/kernel_emitter.cpp
@@ -12,7 +12,7 @@
 
 #include "../test_util/common.hpp"
 #include "gtest/gtest.h"
-#include "nnfusion/core/IR/IR.hpp"
+#include "nnfusion/core/IR/program.hpp"
 #include "nnfusion/core/graph/gnode.hpp"
 #include "nnfusion/engine/profiler/profiler.hpp"
